multiselect: instances and reuse
