+++ include/libp2p/protocol_muxer/multiselect.hpp
@@ -6,19 +6,19 @@
 #ifndef LIBP2P_PROTOCOL_MUXER_MULTISELECT_HPP
 #define LIBP2P_PROTOCOL_MUXER_MULTISELECT_HPP
 
+#include <unordered_set>
+#include <vector>
+
 #include "protocol_muxer.hpp"
-
-#include "multiselect/parser.hpp"
-
-namespace soralog {
-  class Logger;
-}
 
 namespace libp2p::protocol_muxer::multiselect {
 
-  class Multiselect : public protocol_muxer::ProtocolMuxer,
-                      public std::enable_shared_from_this<Multiselect> {
+  class MultiselectInstance;
+
+  class Multiselect : public protocol_muxer::ProtocolMuxer {
    public:
+    using Instance = std::shared_ptr<MultiselectInstance>;
+
     ~Multiselect() override = default;
 
     /// Implements ProtocolMuxer API
@@ -27,92 +27,19 @@ namespace libp2p::protocol_muxer::multiselect {
                      bool is_initiator, bool negotiate_multiselect,
                      ProtocolHandlerFunc cb) override;
 
+    /// Called from instance on close
+    void instanceClosed(Instance instance, const ProtocolHandlerFunc &cb,
+                        outcome::result<peer::Protocol> result);
+
    private:
-    using Protocols = boost::container::small_vector<std::string, 4>;
-    using Packet = std::shared_ptr<MsgBuf>;
-    using Parser = detail::Parser;
-    using MaybeResult = boost::optional<outcome::result<std::string>>;
+    /// Returns instance either from cache or creates a new one
+    Instance getInstance();
 
-    void sendOpening();
+    /// Active instances, keep them here to hold shared ptrs alive
+    std::unordered_set<Instance> active_instances_;
 
-    /// Sends protocol proposal, returns false when all proposals exhausted
-    bool sendProposal();
-
-    /// Sends LS reply message
-    void sendLS();
-
-    /// Sends NA reply message
-    void sendNA();
-
-    void send(MsgBuf msg);
-
-    void send(Packet packet);
-
-    void onDataWritten(outcome::result<size_t> res);
-
-    void close(outcome::result<std::string> result);
-
-    void receive();
-
-    void onDataRead(outcome::result<size_t> res);
-
-    MaybeResult processMessages();
-
-    MaybeResult handleProposal(const std::string_view &protocol);
-
-    MaybeResult handleNA();
-
-    /// Current round, helps enable Multiselect instance reuse (callbacks won't
-    /// be passed to expired destination)
-    size_t current_round_ = 0;
-
-    /// List of protocols
-    Protocols protocols_;
-
-    /// Connection or stream
-    std::shared_ptr<basic::ReadWriter> connection_;
-
-    /// ProtocolMuxer callback
-    ProtocolHandlerFunc callback_;
-
-    /// True for client-side instance
-    bool is_initiator_ = false;
-
-    /// True if multistream protocol version is negotiated (strict mode)
-    bool multistream_negotiated_ = false;
-
-    /// Client specific: true if protocol proposal was sent
-    bool wait_for_protocol_reply_ = false;
-
-    /// True if the dialog is closed, no more callbacks
-    bool closed_ = false;
-
-    /// Client specific: index of last protocol proposal sent
-    size_t current_protocol_ = 0;
-
-    /// Server specific: has value if negotiation was successful and
-    /// the instance waits for write callback completion.
-    /// Inside is index of protocol chosen
-    boost::optional<size_t> wait_for_reply_sent_;
-
-    /// Incoming messages parser
-    Parser parser_;
-
-    /// Read buffer
-    std::shared_ptr<std::array<uint8_t, kMaxMessageSize>> read_buffer_;
-
-    /// Write queue. Still needed because the underlying ReadWriter may not
-    /// support buffered writes
-    std::deque<Packet> write_queue_;
-
-    /// True if waiting for write callback
-    bool is_writing_ = false;
-
-    /// Cache: serialized LS response
-    boost::optional<Packet> ls_response_;
-
-    /// Cache: serialized NA response
-    boost::optional<Packet> na_response_;
+    /// Idle instances which can be reused
+    std::vector<Instance> cache_;
   };
 
 }  // namespace libp2p::protocol_muxer::multiselect
